Get rid of ExternCrate::primitives
